+++ packages/nimble-components/src/rich-text-mention/base/models/mention-validator.ts
@@ -8,7 +8,6 @@ export interface RichTextMentionValidity extends ValidityObject {}
 type FlagNames<T> = T extends readonly (infer U)[] ? U : never;
 
 export const baseValidityFlagNames = [
-    'unsupportedMappingType',
     'duplicateMappingMentionHref',
     'missingMentionHrefValue',
     'mentionHrefNotValidUrl',
@@ -21,12 +20,11 @@ export const baseValidityFlagNames = [
  * Validator for RichTextMention
  */
 export class RichTextMentionValidator<
-    ValidityFlagNames extends readonly string[]
+    ValidityFlagNames extends readonly string[] = typeof baseValidityFlagNames
 > extends Validator<typeof baseValidityFlagNames | ValidityFlagNames> {
     public constructor(
-        private readonly mentionInternals: MentionInternals<unknown>,
-        configValidityKeys: ValidityFlagNames,
-        private readonly supportedMappingElements: readonly (typeof Mapping<unknown>)[]
+        private readonly mentionInternals: MentionInternals,
+        configValidityKeys: ValidityFlagNames
     ) {
         super(configValidityKeys);
     }
@@ -37,7 +35,6 @@ export class RichTextMentionValidator<
     ): void {
         this.untrackAll();
         const mentionHrefs = mappings.map(mapping => mapping.key);
-        this.validateMappingTypes(mappings);
         this.validateNoMissingMentionHref(mentionHrefs);
         this.validateUniqueMentionHref(mentionHrefs);
         this.validateMissingPattern(pattern);
@@ -86,13 +83,6 @@ export class RichTextMentionValidator<
         );
     }
 
-    private validateMappingTypes(mappings: Mapping<unknown>[]): void {
-        const valid = mappings.every(mapping => this.supportedMappingElements.some(
-            mappingClass => mapping instanceof mappingClass
-        ));
-        this.setConditionValue('unsupportedMappingType', !valid);
-    }
-
     private validateUniqueMentionHref(
         mentionHref: (string | undefined | unknown)[]
     ): void {
@@ -126,7 +116,7 @@ export class RichTextMentionValidator<
                 return (
                     href === undefined
                           || typeof href !== 'string'
-                          || !RegExp(pattern).test(href)
+                          || !new RegExp(pattern).test(href)
                 );
             });
         this.setConditionValue('mentionHrefDoesNotMatchPattern', invalid);
@@ -134,6 +124,7 @@ export class RichTextMentionValidator<
 
     private isInvalidUrl(url: string): boolean {
         try {
+            // Check whether the constructor throws an error for an Invalid URL.
             // eslint-disable-next-line no-new
             new URL(url);
             return false;
@@ -144,7 +135,9 @@ export class RichTextMentionValidator<
 
     private isInvalidRegex(pattern: string): boolean {
         try {
-            RegExp(pattern);
+            // Check whether the constructor throws an error when it cannot be parsed as a valid regular expression.
+            // eslint-disable-next-line no-new
+            new RegExp(pattern);
             return false;
         } catch (error) {
             return true;