+++ packages/nimble-components/src/table/components/row/tests/table-row.pageobject.ts
@@ -14,4 +14,18 @@ export class TableRowPageObject<T extends TableRecord = TableRecord> {
             tableCellTag
         )[columnIndex];
     }
+
+    public getRenderedCells(): TableCell[] {
+        return Array.from(
+            this.tableRowElement.shadowRoot!.querySelectorAll<TableCell>(
+                tableCellTag
+            )
+        );
+    }
+
+    public getExpandCollapseButton(): HTMLElement | null {
+        return this.tableRowElement.shadowRoot!.querySelector<HTMLElement>(
+            '.expand-collapse-button'
+        );
+    }
 }