+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,89 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Tue, 28 Nov 2023 18:24:15 GMT and should not be manually modified.
+This log was last generated on Fri, 15 Dec 2023 14:08:18 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 18.2.4
+
+Fri, 15 Dec 2023 14:08:18 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.16.4
+
+## 18.2.3
+
+Wed, 13 Dec 2023 15:42:56 GMT
+
+### Patches
+
+- Add button-label for mention configuration element  ([ni/nimble@28021f0](https://github.com/ni/nimble/commit/28021f000969ab66014f67cbd8a30a175de279f1))
+
+## 18.2.2
+
+Wed, 13 Dec 2023 14:06:08 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.16.3
+
+## 18.2.1
+
+Wed, 13 Dec 2023 13:27:41 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.16.2
+
+## 18.2.0
+
+Mon, 11 Dec 2023 19:32:21 GMT
+
+### Minor changes
+
+- Add contentEditable property to anchor directive ([ni/nimble@487c344](https://github.com/ni/nimble/commit/487c3440588f7b331ea2e90131e41bbf4d4d84f0))
+- Bump @ni/nimble-components to v20.16.1
+
+## 18.1.3
+
+Mon, 11 Dec 2023 17:36:28 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.16.0
+
+## 18.1.2
+
+Fri, 08 Dec 2023 05:30:46 GMT
+
+### Patches
+
+- Angular support for mention configuration and mapping user elements ([ni/nimble@0fef32a](https://github.com/ni/nimble/commit/0fef32ab58b286603588100a5c65906099915d98))
+
+## 18.1.1
+
+Thu, 07 Dec 2023 21:49:16 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.15.1
+
+## 18.1.0
+
+Thu, 07 Dec 2023 17:27:04 GMT
+
+### Minor changes
+
+- Bump @ni/nimble-components to v20.15.0
+
+## 18.0.1
+
+Tue, 05 Dec 2023 11:06:30 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.14.13
 
 ## 18.0.0